feat: adds new AuthModal with sign up form
this update adds a few new components...

Auth - encapsulates both the trigger and the modal into a single
component that can be reused. it allows you to specify what trigger(s)
to show (sign up, sign in, or both), and with any of them are clicked,
will open the AuthModal with the corresponding form displayed.

AuthModal - the modal that will either display the sign up form or the
sign in form, dependong in which method was clicked to open the modal,
or if the user has selected the option within the modal to switch to the
other form.

within this update, the Auth component was added to the global layout as
it will be included in the global header, which will allow the user to
login from any page.

this update only includes the sign up form ui. it does include the sign
in form, or any functionality for signing up. that functionality will
come in later updates.