feat: adds support for opening the auth modal from query params
this update adds support for including the `signin` or `signup` query
params in the url to automatically open the auth modal. if the `signin`
param is found, the auth modal will open with the sign in form
displayed. if the `signup` param is found in the url, the auth modal
will open with the sign up form displayed. this updated functionality
will only execute if the parent component allows it by setting the
`allowOpenFromQueryParams` prop to true.