Issue/12 signup modal UI (#19)
* feat: adds new Button component

this update adds a new Button component which includes configured
stylings based on a `kind` property you can pass to it. using this
configuration (rather than relying on custom CSS in different views and
components) ensures consistency and maintainability throughout the
application.

* feat: adds new TextInput component

this update adds a new TextInput component which encapsulates the common
elements of an input, include the lable, the input itself, additional
text, and an error, as well as styling for the error state. this will
help to ensure consistent layout and styling throughout the application.

* feat: adds new AuthModal with sign up form

this update adds a few new components...

Auth - encapsulates both the trigger and the modal into a single
component that can be reused. it allows you to specify what trigger(s)
to show (sign up, sign in, or both), and with any of them are clicked,
will open the AuthModal with the corresponding form displayed.

AuthModal - the modal that will either display the sign up form or the
sign in form, dependong in which method was clicked to open the modal,
or if the user has selected the option within the modal to switch to the
other form.

within this update, the Auth component was added to the global layout as
it will be included in the global header, which will allow the user to
login from any page.

this update only includes the sign up form ui. it does include the sign
in form, or any functionality for signing up. that functionality will
come in later updates.

* feat: adds support for opening the auth modal from query params

this update adds support for including the `signin` or `signup` query
params in the url to automatically open the auth modal. if the `signin`
param is found, the auth modal will open with the sign in form
displayed. if the `signup` param is found in the url, the auth modal
will open with the sign up form displayed. this updated functionality
will only execute if the parent component allows it by setting the
`allowOpenFromQueryParams` prop to true.

* refactor: prevents btn style transitions on hover/focus when disabled