Issue/27 user session (#31)
* build: installs dayjs and jsonwebtoken

* feat: adds new logger service

only uses console log for now, but will be the central logging service
for the application so can add external logging service later.

* feat: adds JWT util

adds support for signing/creating and reading json web tokens

* feat: adds creating new user session on signup

* feat: adds caching user in redis and retrieving user data with token

* ci: adds redis to test workflow

* ci: adds postgres service container

* refactor: updates setting session cookie so "secure" is only set on prod

there was an issue with some tests failing (integration tests running in
webkit and Mobile Safari) because the cookie was set with the Secure
attribute set to true. within the playwright tests, webkit and Mobile
Safari browsers were failing because the session cookie wasn't being
passed. so this update changes the creation of the cookie so the Secure
attribute is only being set on production.

Additionally, this update cleans up some of the new areas of code to
minimize bloat, and also adds loading the user from the database if
their information isn't found in the cache.

* feat: removes sessions table